Refactor routing and navigation
Wrap router-view in main tag, add new pages, and update AppHeader to use store for menu items.

The routing structure was improved, new pages for projects and about were added, and the AppHeader component was updated to use the store for menu items, ensuring a consistent navigation experience. These changes provide a more organized and scalable approach to routing and navigation within the application.